requirement.txt
fastapi
uvicorn[standard]
langchain
langchain-community
langchain-google-genai
pypdf
faiss-cpu
requests
python-dotenv
gunicorn